moved modules into extras/
I moved requests.js into extras and made the function for finding the
potentional gamepath into it's own module. I also made the exec() called
Promise based.
